Rename enums package
