+++ src/main/java/com/yesthisispatrick/games/minesweeper/models/Board.java
@@ -2,8 +2,8 @@ package com.yesthisispatrick.games.minesweeper.models;
 
 import com.yesthisispatrick.games.minesweeper.services.Configuration;
 import com.yesthisispatrick.games.minesweeper.models.Tile.TileFactory;
-import com.yesthisispatrick.games.minesweeper.enums.COMPASS;
-import com.yesthisispatrick.games.minesweeper.enums.TILE_TYPE;
+import com.yesthisispatrick.games.minesweeper.constants.COMPASS;
+import com.yesthisispatrick.games.minesweeper.constants.TILE_TYPE;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
