+++ src/test/java/com/yesthisispatrick/games/minesweeper/constants/TileTypeTest.java
@@ -1,8 +1,7 @@
-package com.yesthisispatrick.games.minesweeper.enums;
+package com.yesthisispatrick.games.minesweeper.constants;
 
 import static org.junit.Assert.*;
 
-import com.yesthisispatrick.games.minesweeper.enums.TILE_TYPE;
 import org.junit.Test;
 
 public class TileTypeTest {
