+++ src/test/java/com/yesthisispatrick/games/minesweeper/models/TileTest.java
@@ -2,9 +2,8 @@ package com.yesthisispatrick.games.minesweeper.models;
 
 import static org.junit.Assert.*;
 
-import com.yesthisispatrick.games.minesweeper.models.Tile;
 import com.yesthisispatrick.games.minesweeper.models.Tile.TileFactory;
-import com.yesthisispatrick.games.minesweeper.enums.TILE_TYPE;
+import com.yesthisispatrick.games.minesweeper.constants.TILE_TYPE;
 import org.junit.Before;
 import org.junit.Test;
